backend/requirements.txt
fastapi
uvicorn
google-generativeai
psycopg2-binary
pydantic
python-dotenv